+++ R/visualize_game.R
@@ -8,18 +8,24 @@
 #'                 if \code{FALSE} don't annotate,
 #'                 if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.
 #' @param ... Arguments to \code{pmap_piece}
+#' @param .f Low level graphics function to use e.g. \code{grid.piece}, \code{piece3d}, or \code{piece}.
 #' @param cfg A piecepackr configuration list
 #' @param envir Environment (or named list) of piecepackr configuration lists
 #' @param n_transitions Integer, if over zero (the default)
 #'                how many transition frames to add between moves.
 #' @param n_pauses Integer, how many paused frames per completed move.
 #' @param fps Double, frames per second.
+#' @param width Width of animation (in inches).  Inferred by default.
+#' @param height Height of animation (in inches).  Inferred by default.
+#' @param ppi Resolution of animation in pixels per inch.
+#'            By default set so image max 600 pixels wide or tall.
 #' @param new_device If \code{file} is \code{NULL} should we open up a new graphics device?
 #' @return Nothing, as a side effect saves an animation of ppn game
 #' @export
 animate_game <- function(game, file = "animation.gif", annotate = TRUE, ...,
-                         cfg = NULL, envir = NULL,
+                         .f = piecepackr::grid.piece, cfg = NULL, envir = NULL,
                          n_transitions = 0L, n_pauses = 1L, fps = n_transitions + n_pauses,
+                         width = NULL, height = NULL, ppi = NULL,
                          new_device = TRUE) {
 
     if (n_transitions > 0L && !requireNamespace("tweenr", quietly = TRUE)) {
@@ -43,33 +49,19 @@ animate_game <- function(game, file = "animation.gif", annotate = TRUE, ...,
     ymin_op <- min(sapply(ranges, function(x) x$ymin_op), na.rm = TRUE)
     xmax <- max(sapply(ranges, function(x) x$xmax), na.rm = TRUE)
     ymax <- max(sapply(ranges, function(x) x$ymax), na.rm = TRUE)
-    if (xmin_op < 0.50) {
-        xoffset <- 0.50 - xmin_op
-    } else {
-        xoffset <- 0
-    }
-    if (ymin_op < 0.50) {
-        yoffset <- 0.50 - ymin_op
-    } else {
-        yoffset <- 0
-    }
-    width <- xmax_op + xoffset + 0.50
-    height <- ymax_op + yoffset + 0.50
+    xoffset <- min2offset(xmin_op)
+    yoffset <- min2offset(ymin_op)
+    if (is.null(width)) width <- xmax_op + xoffset + 0.50
+    if (is.null(height)) height <- ymax_op + yoffset + 0.50
     m <- max(width, height)
-    res <- round(600 / m, 0)
+    if (is.null(ppi)) ppi <- round(600 / m, 0)
     # mp4 needs even height / weight
-    height <- ceiling(res * height)
-    width <- ceiling(res * width)
+    height <- ceiling(ppi * height)
+    width <- ceiling(ppi * width)
     height <- height + (height %% 2)
     width <- width + (width %% 2)
-    plot_fn <- function(df, ...) {
-        df$x <- df$x + xoffset
-        df$y <- df$y + yoffset
-        grid::grid.newpage()
-        pmap_piece(df, default.units = "in", ..., envir = envir)
-        annotate_plot(annotate, xmax, ymax, xoffset, yoffset)
-    }
-    animation_fn(file, new_device)(lapply(dfs, plot_fn, ...), file, width, height, 1 / fps, res)
+    plot_fn <- plot_fn_helper(.f, xmax, ymax, xoffset, yoffset, width, height, m, ppi, envir, annotate)
+    animation_fn(file, new_device)(lapply(dfs, plot_fn, ...), file, width, height, 1 / fps, ppi)
     invisible(NULL)
 }
 #### How to handle empty tibbles??
@@ -84,7 +76,7 @@ animation_fn <- function(file, new_device = TRUE) {
     } else if (grepl(".html$", file)) {
         if (!requireNamespace("animation")) stop("You need to install the suggested package 'animation'")
         function(expr, file, width, height, delay, res) {
-            animation::saveHTML(expr, htmlfile = file, interval = delay,
+            animation::saveHTML(expr, htmlfile = file, interval = delay, img.name = file,
                                 ani.height = height, ani.width = width, ani.res = res,
                                 ani.dev = "png", ani.type = "png",
                                 title = "Animated game", verbose = FALSE)
@@ -96,7 +88,7 @@ animation_fn <- function(file, new_device = TRUE) {
             }
         } else if (requireNamespace("animation", quietly = TRUE)) {
             function(expr, file, width, height, delay, res) {
-                animation::saveGIF(expr, movie.name = file, interval = delay,
+                animation::saveGIF(expr, movie.name = file, interval = delay, img.name = file,
                                    ani.height = height, ani.width = width, ani.res = res,
                                    ani.dev = "png", ani.type = "png")
             }
@@ -107,7 +99,7 @@ animation_fn <- function(file, new_device = TRUE) {
     } else {
         if (!requireNamespace("animation")) stop("You need to install the suggested package 'animation'")
         function(expr, file, width, height, delay, res) {
-            animation::saveVideo(expr, video.name = file, interval = delay,
+            animation::saveVideo(expr, video.name = file, interval = delay, img.name = file,
                                 ani.height = height, ani.width = width, ani.res = res,
                                 ani.dev = "png", ani.type = "png",
                                 title = "Animated game", verbose = FALSE)
@@ -174,28 +166,68 @@ get_df_from_move <- function(game, move = NULL) {
     }
 }
 
+plot_fn_helper <- function(.f = grid.piece, xmax, ymax, xoffset, yoffset,
+                           width, height, m, ppi, envir, annotate) {
+    if (identical(.f, grid.piece)) {
+        function(df, ..., scale = 1) {
+            df$x <- df$x + xoffset
+            df$y <- df$y + yoffset
+            df$scale <- if (has_name(df, "scale")) scale * df$scale else scale
+            grid::grid.newpage()
+            pmap_piece(df, default.units = "in", ..., envir = envir)
+            annotate_plot(annotate, xmax, ymax, xoffset, yoffset)
+        }
+    } else if (identical(.f, piece3d)) {
+        if (!requireNamespace("rgl")) stop("You need to install the suggested package 'rgl'")
+        if (Sys.which("wmctrl") != "") system(paste0("wmctrl -r RGL -e 0,-1,-1,", width, ",", height))
+        f <- tempfile(fileext=".png")
+        function(df, ..., scale = 1) {
+            df$scale <- if (has_name(df, "scale")) scale * df$scale else scale
+            rgl::rgl.clear()
+            pmap_piece(df, piece3d, ..., envir = envir)
+            Sys.sleep(2)
+            rgl::rgl.snapshot(f, top = FALSE)
+            grid::grid.newpage()
+            grid::grid.raster(png::readPNG(f))
+        }
+    } else if (identical(.f, piece)) {
+        if (!requireNamespace("rayrender")) stop("You need to install the suggested package 'rayrender'")
+        function(df, ..., scale = 1) {
+            df$scale <- if (has_name(df, "scale")) scale * df$scale else scale
+            df$x <- df$x + xoffset
+            df$y <- df$y + yoffset
+            l <- pmap_piece(df, piece, ..., envir = envir)
+            table <- rayrender::sphere(z=-1e3, radius=1e3, material=rayrender::diffuse(color="green"))
+            light <- rayrender::sphere(x=0.5*width/ppi, y=-4, z=max(1.5*m+1, 20),
+                                       material=rayrender::light(intensity=420))
+            table <- rayrender::add_object(table, light)
+            scene <- Reduce(rayrender::add_object, l, init=table)
+            rayrender::render_scene(scene,
+                                    lookat = c(0.5*width/ppi, 0.5*height/ppi, 0),
+                                    lookfrom = c(0.5*width/ppi, -7, 1.5*m),
+                                    width = width, height = height, samples=200)
+        }
+    } else {
+        .f
+    }
+
+
+}
+
 #' Plot game move
 #'
 #' Plot game move
-#' @param game A list containing a parsed ppn game (as parsed by \code{read_ppn})
-#' @param file Filename to save graphic to unless \code{NULL}
-#'             in which case it uses the current graphics device.
+#' @inheritParams animate_game
 #' @param move Which move to plot game state (after the move, will use \code{game$dfs[[move]]})
 #'             unless \code{NULL} in which case will plot the game state after the last move.
-#' @param annotate If \code{TRUE} annotate the plot with \dQuote{algrebraic} coordinates,
-#'                 if \code{FALSE} don't annotate,
-#'                 if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.
 #' @param bg Background color (\code{"transparent")} for transparent
-#' @param res For bitmap image formats the resolution
-#' @param ... Arguments to \code{pmap_piece}
-#' @param cfg A piecepackr configuration list
-#' @param envir Environment (or named list) of piecepackr configuration lists
-#' @param new_device If \code{file} is \code{NULL} should we open up a new graphics device?
 #' @return Nothing, as a side effect saves a graphic
 #' @import grDevices
 #' @export
-plot_move <- function(game, file = NULL,  move = NULL, annotate = TRUE, ..., bg = "white", res = 72,
-                      cfg = NULL, envir = NULL, new_device = TRUE) {
+plot_move <- function(game, file = NULL,  move = NULL, annotate = TRUE, ...,
+                      .f = piecepackr::grid.piece, cfg = NULL, envir = NULL,
+                      width = NULL, height = NULL, ppi = 72,
+                      bg = "white",  new_device = TRUE) {
 
     ce <- piecepackr:::default_cfg_envir(cfg, envir)
     cfg <- ce$cfg
@@ -203,38 +235,40 @@ plot_move <- function(game, file = NULL,  move = NULL, annotate = TRUE, ..., bg 
 
     df <- get_df_from_move(game, move)
     dfr <- range_true(df, cfg = cfg, envir = envir, ...)
-    if (dfr$xmin_op < 0.50) {
-        xoffset <- 0.50 - dfr$xmin_op
-    } else {
-        xoffset <- 0
-    }
-    if (dfr$ymin_op < 0.50) {
-        yoffset <- 0.50 - dfr$ymin_op
-    } else {
-        yoffset <- 0
-    }
-    width <- dfr$xmax_op + xoffset + 0.50
-    height <- dfr$ymax_op + yoffset + 0.50
-    df$x <- df$x + xoffset
-    df$y <- df$y + yoffset
-
+    xmax <- dfr$xmax
+    ymax <- dfr$ymax
+    xoffset <- min2offset(dfr$xmin_op)
+    yoffset <- min2offset(dfr$ymin_op)
+    if (is.null(width)) width <- dfr$xmax_op + xoffset + 0.50
+    if (is.null(height)) height <- dfr$ymax_op + yoffset + 0.50
+    m <- max(width, height)
     if (is.null(file)) {
         if (new_device) dev.new(width = width, height = height, unit = "in", noRstudioGD = TRUE)
     } else {
         format <- tools::file_ext(file)
         switch(format,
-               bmp = bmp(file, width, height, "in", res = res, bg = bg),
-               jpeg = jpeg(file, width, height, "in", res = res, bg = bg),
+               bmp = bmp(file, width, height, "in", res = ppi, bg = bg),
+               jpeg = jpeg(file, width, height, "in", res = ppi, bg = bg),
                pdf = cairo_pdf(file, width, height, bg = bg),
-               png = png(file, width, height, "in", res = res, bg = bg),
+               png = png(file, width, height, "in", res = ppi, bg = bg),
                ps = cairo_ps(file, width, height, bg = bg),
                svg = svg(file, width, height, bg = bg),
-               tiff = tiff(file, width, height, "in", res = res, bg = bg))
+               tiff = tiff(file, width, height, "in", res = ppi, bg = bg))
     }
-    pmap_piece(df, default.units = "in", ..., envir = envir)
-    annotate_plot(annotate, dfr$xmax, dfr$ymax, xoffset, yoffset)
+    # plot_fn_helper expected width and height in pixels
+    width <- ppi * width
+    height <- ppi * height
+    plot_fn_helper(.f, xmax, ymax, xoffset, yoffset, width, height, m, ppi, envir, annotate)(df, ...)
     if (!is.null(file)) dev.off()
     invisible(NULL)
+}
+
+min2offset <- function(min) {
+    if (min < 0.50) {
+        0.50 - min
+    } else {
+        0
+    }
 }
 
 annotate_plot <- function(annotate, xmax, ymax, xoffset = 0, yoffset = 0) {